requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.1
discord.py==2.2.3
emoji==1.2.0
emoji-country-flag==1.2.4
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
gTTS==2.3.2
idna==2.10
multidict==5.1.0
packaging==20.9
pipdate==0.5.3
pkg-resources==0.0.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytz==2021.1
requests==2.28.2
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
